Tests : Don't write to `/tmp`
Use `self.temporaryDirectory()` instead, which doesn't assume the existence of any particular file structure, and is cleaned up automatically.

I have simply removed `ReferenceTest.testLoadScriptWithReferenceFromVersion0_14` because keeping it would be a bit tricky - it uses a file with a hardcoded reference to `/tmp` in it. I think we're OK to lose the coverage at this point - it's testing compatibility with files from 2015.