ScaleHandle : Normalize uniform `scaling()`
`updatePosition()` returns coordinates in gadget-space, which does not include scaling based on the viewport position. This works fine for non-uniform scaling modes because they are normalized by the start position, which is also high when zoomed out, low when zoomed in.

We use `rasterScaleFactor()` to normalize the scaling, making `scaling()` more consistent with the gadget's viewport size and prevents very large scaling when zoomed out, very small scaling when zoomed in. The `m_startPosition()` subtraction is dropped because it is insignificant compared to `rasterScaleFactor()`.