PointConstraint : Move `enabled[XYZ]` plugs above `offset` plug
What I really wanted to do was to also add UI activators to `offset.[xyz]` so that each component was only active when the appropriate toggle was on. But that didn't work since CompoundNumericPlugValueWidget doesn't use a PlugLayout internally. But I still think this new order is _slightly_ better, in that reading top-to-bottom you don't consider `offset` until you know which components are going to be relevant.