PathColumn : Add signals to allow custom button event handling
This will allow us to encapsulate the event handling for a column along with the code for generating the cell data, allowing non-intrusive interactive extensions to things like the LightEditor and HierarchyView. The immediate use cases are custom columns for handling mute/solo and visible set management respectively.

The PathListingWidgetAccessor code is somewhat convoluted, but allows C++ PathColumns to have sufficient access to the PathListingWidget state while also allowing Python PathColumns full access using the exact same signals. The only realistic alternative I could see was only providing the signals in Python, but we want to encourage C++ PathColumns because they are typically quite performance-critical. In a far-distant future I would love to unify Gadget and Widget so that there is a C++ base class for both (or they become one and the same thing). That would allow a slightly more natural implementation.