CatalogueUI : Use PathColumn event handling for output index column
This allows us to encapsulate all the handling for the output index into the column class, rather than having display code in one place and editing scattered through another class. This will become increasingly important as we add similar custom columns to the LightEditor and HierarchyView in future.

Also added `__` prefix to `__OutputIndexColumn`, as it wasn't intended to be public.