FileSystemPath : Handle escaped characters
Because Windows can use a backslash to separate paths, and Gaffer can use a backslash to indicate a literal `$`, `#` or `~`, we need a way of indicating what a backslash is used for.

Given a `\$`, `\#` or `\~`, we settle on treating it as an escaped special character as it would be on Linux. To prevent boost from treating it as a path separator, we substitute a temporary value (an illegal file path character so we don't have to worry about incorrectly substituting intended characters) before having boost decompose the path.

Those standin characters are then reversed when setting the root and name values.

One case is left, which is a special character preceding what is meant to be a directory separator. In this case, the user must use a double backslash - one for the directory separator, one for the escaped character. Or use forward slashes throughout, which is what we do in `FileSystemPath::doChildren()`.