PathListingWidgetTest : Check exact expected expansion depth
The `_pathListingWidgetAttachTester()` call had convinced me we had a bug where the PathModel merrily visited children it hadn't been asked to visit. Fortunately, we didn't have a bug, and the tester was just visiting additional children of its own accord. Strengthened the test to verify that we don't visit any unexpected children.