OSLCode : Fix compilation with Boost 1.73
We don't officially support this Boost version, but Image Engine is using it for some reason. Officially we are using Boost 1.76 from VFX Reference Platform 2022.

We're also not sure if changing from `GraphComponentPtr` to `GraphComponent *` would be an ABI break, so we're playing it safe and only making the change for builds with old Boost.