Capsule : Remove tracking of `m_scene` parent
This was intended to provide some safety rails to diagnose the use of invalidated capsules. But in practice, the reparenting of a ScenePlug is just one of a great many ways of invalidating a capsule, none of which we have protected against since 02ac22ee8a5ea91716f3f9bb6f714198515d5648.

My main motivation here was actually to remove one more use of `Signal::disconnect( slot )`, so that we can eventually remove that method. It would be perfectly possible to keep the parent tracking by using a `ScopedConnection` member variable to do the disconnection with instead. But given the documented thread-safety issues with using signals in this context, and the general ineffectiveness of the invalidation tracking, I thought it preferable just to remove it.

Note that `throwIfNoScene()` still has a use - it detects the case where the scene was lost during save/load of a Capsule.