IECoreCycles : Replace ObjectAlgo with GeometryAlgo
None of the converters had any need to control the creation of the object - geometry is the only thing they care about, so it is the only thing they should return. This allows us to remove the insane circular logic in InstancCache, where `InstanceCache::get()` was calling `InstanceCache::convert()` even for a _cache hit_.

This also deals correctly with null returns from `GeometryAlgo::convert()`, which fixes #4865.