IECoreCycles : Fix crash rendering PointsPrimitives with `N`
We were attempting to identify "standard" attributes and use the `AttributeSet::add( AttributeStandard )` method for creating them. This had at least two problems :

- We were only testing the primitive variable name, and not making any attempt to verify that the data type and interpolation matched.
- `AttributeSet::add()` crashes when trying to create `ATTR_STD_VERTEX_NORMAL` for points geometry, and various other combinations.

Instead, we now create a generic attribute by the normal method, and afterwards tag it as standard if it meets the relevant criteria. It may be that we _shouldn't_ be sidestepping the `AttributeSet` crash like this, and that there's a very good reason not to add `ATTR_STD_VERTEX_NORMAL` to point clouds. But the unit test demonstrates that it is actually necessary to do so if the `attribute` shader is to be able to pick up `N` (it can pick up non-standard names without special treatment, but _needs_ `Attribute::std` to be set to pick up `N`).

The indentation has changed on the "Create attribute" section, but the logic for generic attribute creation is unchanged.

I removed the creation of standard attributes for `uv` and `uTangent` because I don't understand why UVs were being dealt with here when MeshAlgo has a completely custom code path for those anyway. Instead I've added a todo to investigate. It might be that we can refactor to avoid a bunch of code duplication. And it might be that it should be the specific geometry algo that should apply the standard tags. To be continued...

Fixes #4876.