Cycles RendererTest : Fix pixel indexing
If there are an even number of rows, then the middle pixel is not `numPixels // 2`. This reveals an unfortunate fact : Cycles has "light leakage" on the triangle seam in the middle of our plane. We work around that for now by slightly offsetting the test position, since these particular tests are testing other things.