GafferScene : Add VisibleSet
This introduces a simple struct used to define a subset of the scene
hierarchy to be rendered. This allows for:
`expansions` supporting the usual UI based approach of interactively
expanding scene locations in order to render them.
`inclusions` which are locations that are always rendered.
`exclusions` which are locations that are never rendered.