RenderController : Update to use VisibleSet
Rather than use a single PathMatcher containing the expanded locations,
we now use a VisibleSet which provides `inclusions` and `exclusions`
in addition to the usual `expansions`.

These inclusions and exclusions inform us of locations which should
always be rendered, or should never be rendered, despite the current
contents of expansions. Additionally, exclusions allow us to not render
locations despite them falling within the minimumExpansionDepth.

As the current expansion behaviour requires a location and all of its
ancestors to be expanded before that location will be rendered, any
location added to inclusions needs to also have all of its ancestors
expanded before it can render.  There's a todo here to revisit this
behaviour for inclusions. Removing the need for expansion of an
inclusion's ancestors would provide advantages such as not requiring
the sibling locations of an inclusion to be rendered if they are
not also present in expansions or inclusions.