Gadget : Avoid `Signal::disconnect( slot )`
Currently we rarely give a specific style to an individual Gadget, because we prefer to use metadata to control the appearance of nodes and the like. But once upon a time we used to control the appearance of nodes by giving them a style. That would have been the worst case scenario for `Signal::disconnect()`, with lots of Gadgets all having a connection to the same signal, and all paying linear time for the disconnect. We're planning to remove `Signal::disconnect()` to avoid this sort of pitfall, so we now use explicit connections with constant-time `disconnect()`.

Since f36d27b005e11a4638fd0d376c225e748b0e9bd7 showed that draw time was quite sensitive to `sizeof( Gadget )`, I also tried an alternative implementation where both the StylePtr and the connection were held indirectly in a separate struct, keeping `sizeof( Gadget )` as it currently is. If anything, there was _maybe_ ~2% improvement in drawtime with that vs this commit, but the test was pretty noisy and I deemed it not worth the small cost in extra complexity.