IECoreCycles : Fix primitive variable interpolation bugs
The rules for converting from `PrimitiveVariable::Interpolation` to `ccl::AttributeElement` are subtle and specific to each type of geometry, so we make that the responsibility of the individual type-specific converters rather than trying to deal with it in AttributeAlgo.

Among other things, this fixes the handling of `vertex Cd` in #4849. In fact, that issue was using `Varying` interpolation, which while it should have been equivalent to `Vertex` was instead being treated as equivalent to `FaceVarying`.