Cycles MeshAlgo : Remove normals code
This means we handle normals using the standard `convertPrimitiveVariables()` code path, and facevarying normals no longer cause crashes. In theory we might be losing something from the handling of multiple normals for motion blur, but in practice motion blur isn't working properly anyway. Better to strip this back to the basics and rebuild...

Fixes #4849.