Cycles MeshAlgo : Remove custom UV conversion code
The standard primitive variable conversion code can convert the data just fine. Our convention is that the standard uv set is called "uv", so we just need to tag that as standard and we're done. None of "st_0", "st0" etc have any claim on being the standard UV set, so we just treat them as custom primitive variables.