Cycles MeshAlgo : Remove support for `_smooth` primitive variable
This is nonstandard, and we don't have any ways of authoring it meaningfully. It would be good to have better support for smooth/hard normals, but our standard way of representing them is with FaceVarying interpolation, and Cycles doesn't support that. See https://developer.blender.org/D12902 for a proposal to add support.