Cycles MeshAlgo : Remove "support" for `_facesetIndex`
My best guess is that the intention behind this was to allow shader assignments per-face, with the face set index being an index into a list of shaders on the mesh. But we never assign such a list of shaders in Renderer.cpp, and we have no convention for even defining what they would be in Gaffer.