ImageView : Minor fix for CPU color path
We're planning to deprecate this ( and maybe it should be more deprecated already ... is it weird that the Alt+G hotkey still exists as a hidden feature, even though we've removed the corresponding UI since it's incompatible with wipes? ).

Anyway, in the meantime, as long as this code exists, it's better if the clipping doesn't default on, and require you to click it on and off again to get back to the correct default value.