VisibleSet : Clarify the result of `match`
We now return an ExactMatch for any location considered visible by
the VisibleSet, and a DescendantMatch for locations with visible
descendants.

The RenderController then needs to only consider these two results
when determining whether to expand a location. We still require the
DescendantMatch to ensure that locations added to a VisibleSet's
`inclusions` have their ancestors expanded in order to be visible.