PathModel : Improve expansion handling
The old logic dates from a time when we had to create the Qt indices for _all_ expanded paths, even if their ancestors weren't expanded. That was necessary because the Qt expansion state was our source of truth, and `getExpansion()` would read from it. That's not necessary any more, because the PathMatcher itself is our source of truth.

There's no need to eagerly create children that won't be visible in the QTreeView because their parent isn't expanded. But where the parent _is_ expanded we can avoid a second update loop by creating the children before we expand the parent in Qt (which will cause Qt to query the children).