SceneReader/SceneWriter : Prefer SceneInterface sets API for USD & ABC
This is a better match for the data structures in Gaffer (PathMatcher), USD (UsdCollection) and Alembic (AbcCollection), resulting in simpler implementation and improved performance. This gives a greater than 2x speedup reading a large set from USD. It _should_ also have meant that we supported the writing of sets to Alembic files, but that won't be the case until IECoreAlembic is fixed to support writing sets at the root of the scene.

A side effect of this is that cancellation is now responsive when reading large sets from USD. There is no cancellation API for `readTags()`, and USDScene was emulating tags by loading the entire set on the first call to `readTags()`.

I've implemented this as a blacklist of formats that use tags rather than a whitelist of formats that use sets, as going forwards I think all new formats should prefer to implement the sets API.