PythonExpressionEngine : Fix non-deterministic plug parse order
We store plug paths in a set as we encounter them (to avoid duplicates), and then turn the set into a list. But in Python 3, the order of iteration of a set is non-deterministic from process-to-process [^1], so the parser was not providing repeatable results. Sorting the results before returning them fixes this.

This was a potential fix for #4935, which in the end was addressed more satisfactorily by #4943. But a deterministic parse order still seems like something we should have.

[^1]: Sets are ordered according to the hash of their members, and Python 3's string hash is "salted" with a different random value per process. See https://docs.python.org/3/using/cmdline.html#cmdoption-R.