PlugValueWidget : Fix unnecessary updates from `setContext()`
In an ideal world, Python's `is` would work as the expected identity test for `IECore.RefCounted` objects just as `Gaffer.Context`. But it doesn't, because the Cortex bindings can't guarantee not to make multiple Python objects referring to the same RefCounted object. So we must use the `isSame()` workaround provided by Cortex.

This fixes a double update when displaying a PlugValueWidget in the NodeEditor for the first time.