NodeToolbar/PlugLayout : Fix unnecessary updates
These both had a fallback context that was different to PlugValueWidget's, and didn't have any concept of the appropriate context when showing plugs from a View. This led to no less than 4 updates per PlugValueWidget when creating the Viewer toolbars for the first time! Only the final update had the correct context from the ScriptNode, with all the others using a fallback context.

This problem went unnoticed until now because most toolbar widgets weren't context-sensitive. But the recent addition of the `ImageView.view` plug changed that, because ViewPlugValueWidget uses the context to compute the `viewNames` for the image being viewed.