Arnold Renderer : Support InternedStringData for user attributes
We wouldn't expect people to be authoring this in Gaffer, but we're seeing it being loaded from USD files containing primvars of type `token`. My understanding is that `token` is only intended for use for "enum" type primvars and `string` should be understanding for open-ended values, but that doesn't seem to always be how things are being done in the wild.