Spreadsheet : Improve interoperability with SetQuery
This allows for an elegant setup where the rows of the spreadsheet contain the names of sets, `Spreadsheet.enabledRowNames` drives `SetQuery.sets`, and `SetQuery.firstMatch` drives `Spreadsheet.selector`. This allows the spreadsheet to output different values depending on the first set a location is a member of. And adding new rows automatically updates the SetQuery to include those sets in the query.

The original Spreadsheet implementation prevented this by introducing a circular dependency between `selector -> enabledRowNames -> firstMatch`. But this dependency was only necessary as an optimisation for cases where `selector` contained `${variable}` substitutions. We know that is not the case here, so can bypass the optimisation. The code isn't pretty, but the node networks enabled by it are.