PathListingWidget : Fix tracking of last selected index
QModelIndex must _never_ be stored, as it contains a pointer directly into items in the model, that could be deleted at any time. Instead we must use QPersistentModelIndex, which allows us to query the validity of the index before retrieving it. We were doing that originally, but QModelIndex crept back in when we were adding per-column selection handling.

One reliable way of exercising this bug was as follows :

1. Load a Catalogue containing 3 images into the NodeEditor.
2. Click to select the 2nd image.
3. Right click on the header, and add a column. This resets the model, invalidating all previous indexes.
4. Making sure the mouse has never left the image listing (so we don't lose keyboard focus), press <kbd>↓</kbd> to try to move the selection down. This should crash.