Wrapper : Set `ARNOLD_ADP_OPTIN=0` if not set
This disables the Autodesk Analytics program, unless someone wants to opt-in by setting it to something else before running Gaffer.

Historically, ADP has caused various problems when using Arnold in a studio environment : hangs, out-of-disk-space errors, failed attempts to "phone home" from behind a firewall etc. This is meant to be improved in the latest Arnold release, but that introduces a new problem : Arnold now pops up a dialogue asking people to opt-in to ADP, and it'll try to do so even on a _headless_ machine. It is our responsibility to avoid the problems caused by that by calling `AiADPDisableDialogWindow()`, and we could probably do that in a config file for `gaffer execute` to cover most headless use cases. But even in a GUI session, the ADP window has problems : Gaffer hangs while it is open and the window doesn't stay on top, so it's easy to lose it and think Gaffer is dead. And on my machine at least, the layout is out of whack and the text is clipped.

So, we could just _always_ call `AiADPDisableDialogWindow()`, and potentially provide a menu item to call `AiADPDisplayDialogWindow()` so people can open it if and when they want to. But instead I've gone all grumpy-old-man and just disabled ADP entirely; it doesn't have the same quality of implementation as Arnold, phoning home isn't a good fit for a studio environment, and commercial analytics don't seem a good default behaviour for an open-source product. I apologise to the sincere Arnold developers who would like to use it to improve the product, but folks can still opt-in using the environment variable if they want to.