FileSequencePathFilter : Ignore numbers in directory names
Here we're really covering for a bug in Cortex, where numbers in directory names are being considered to be sequence indicators, rather than limiting it to numbers in the leaf filename. As discussed in https://github.com/GafferHQ/gaffer/pull/4869, we do want to fix this in Cortex eventually, but for now I'm just working around it here.