Avoid deadlocks when `computeCachePolicy()` requires mutex
The issue here is the one documented in the new comments in `ComputeNode::computeCachePolicy()`. The `Standard`, `TaskCollaboration` and `TaskIsolation` modes all hold a lock on the compute cache entry while the compute is taking place. If an upstream compute with the _same hash_ is triggered during this compute, then we would deadlock trying to reacquire the lock (either from this thread or another). So we must avoid that by making sure the hash is unique somehow.

Things are a little more nuanced than that though. For a long time now we've had a special case for `TaskCollaboration` that we thought allowed for this kind of recursion. This is the `TaskMutex::WorkerRead` lock type, whereby a worker thread already doing work for `TaskMutex::execute()` is allowed to re-acquire the lock recursively. We even had coverage for this in `SetTest.testRecursion()` and `FilterResults.testComputeCacheRecursion()`. But recently we hit deadlocks in a production setup where this wasn't working. Although we do have a simple interactive repro for this, I haven't been able to distill it down into a unit test. But I _believe_ the gist of what was happening is this :

    ComplexScene
	     |
      ParentA (filter not matching anything, task arena X)
	     |
	  ParentB (filter matching something, task arena Y)
         |
	  ParentC (filter not matching anything, task arena X)

The hash for `ParentA.__branches` was the same as the hash for `ParentC.__branches`, so they shared the _same_ cache entry. But the intervening ParentB had a different cache entry. Taking into account evictions, a compute for ParentA could be the holder of the mutex, with a compute for ParentC doing its own work in the same arena via a WorkerRead lock. Or ParentC could be the holder of the mutex, with a ParentA compute happening via a WorkerRead lock. Except in this latter case, the ParentA tasks are going into ParentB's arena, due to a deficiency in our LRUCache logic. This is all too tricky for me to reason about, but I think somewhere therein lies the reason for deadlock.

While trying to add a repro to the unit tests, I came up with another test which was equally scary : `ParentTest.testChainedNodesWithIdenticalBranches()`. This is as above, but omitting the central node. What this test shows is that a single cache eviction can cause worker recursion to lead to 1000s of repeat computes. We start a compute for `ParentC.__branches`, which makes multiple upstream tasks which all end up pulling on `ParentA.__branches`. Because these hit the same cache entry, they each are allowed in with WorkerRead privileges, and they each launch their own _duplicate_ compute for `ParentA.__branches`.

This second discovery lead me to believe that the entire WorkerRead concept was flawed, and it was not worth trying to fix the deadlock. The fix I've made instead is to document that hashes must be such that compute cache recursion is impossible, and to uniquefy the hashes for nodes violating this requirement.

This fix gives a 1000x speedup in `testChainedNodesWithIdenticalBranches()`.