ValuePlug : Avoid deadlock between TaskParallel and Legacy compute
The new MeshToLevelSet test models a situation where a downstream TaskCollaboration compute triggers an upstream Legacy compute with the _same_ hash. In this case the TaskCollaboration is on `meshToLevelSet2.__processedObject` and the Legacy compute is on `meshToLevelSet1.out.object`. We were calling `cache.set()` to store the result of the upstream compute, but this failed because it tried to get a lock, and the cache entry is already locked by the downstream compute. The new `setIfUncached()` method is just a no-op in this case.

The new Rename test demonstrates an even simpler way of triggering the problem, by implementing a TaskCollaboration compute to perform a pass-through. Arguably the expensive TaskCollaboration should have been performed on an internal plug and delegated to by `out.set`, like `ObjectProcessor.out.object` delegates to `ObjectProcessor.__processedObject`. But it's pleasing that this fix helps in this case too.

There aren't any unit tests to cover the `setIfUncached()` calls now used in the `Process::forceMonitoring()` case, but I think it is logical that we need the same here to avoid a theoretical deadlock, and I can't see any downside to using them.

I've actually been meaning to add `setIfUncached()` anyway, because it allows us to avoid two cache lookups, one for `get()` and one for `set()`. So this is something we would be doing even without discovering the deadlock.