Serialisation APIs : Use `std::filesystem::path` to represent paths
This includes the methods of ApplicationRoot, Box, Reference and ScriptNode.

We use `filesystem::path` in C++ and convert to and from `pathlib.Path` automatically in the Python bindings. Pleasingly, this allows us to drop a whole load of `.as_posix()` calls from the Python tests.

One annoyance is that there is no value for an empty `pathlib.Path`, so when converting an empty `filesystem::path` to Python we return `None`. This is logical given the constraints of the `pathlib` API, but it does mean that callers will need to be careful to handle the two different return types as necessary.