GafferDispatch : Adapt to `filesystem::path` and `pathlib.Path`
There is one API change here, where `Dispatcher.jobDirectory()` now returns a `filesystem::path`. The rest is just updating the unit tests to account for this and the previous change of return type for `self.temporaryDirectory()`.

Rather than make the absolute minimum of changes to the unit tests, I've tried to use modern Python features where it made sense :

- Prefer `pathlib.Path` methods over `os.path` and `glob` functions, even though the latter do accept `Path` arguments.
- Use f-strings for formatting, instead of string concatenation.