ie/options : Avoid use of list comprehension in config file
In Python 3 list comprehensions were changed to create their own local
scope. If we pass in different objects for the globals and locals to the
exec function this leads to NameErrors in the list comprehension. SCons
passes in two different dictionaries when calling exec on the options
files. To circumvent this issue, we rewrite the list comprehension to a
loop.

For reference:
https://stackoverflow.com/questions/45132645/list-comprehension-in-exec-with-empty-locals-nameerror