Inspector : Call `Inspector::editFunction()` only once
We were calling it every time the history exited the EditScope, and clobbering our stored edit function each time. `AttributeInspectorTest.testDontEditParentOfInspectedLocation` demonstrates a failure caused by the clobbering : because the internal AttributeTweaks node has `localise = True`, it evaluates the upstream graph for every ancestor location as well, with the history for the most distant ancestor being the one that exits the EditScope last. We were therefore preferring to edit ancestor locations instead of the true location.

The fix is cheap and cheerful : just use the first branch that exits the EditScope, and ignore all others. In practice, this works pretty well because most SceneProcessor nodes pull on their primary input at the current location before they pull on secondary inputs and locations. A more principled fix might be to consider the first branch where `scene:path` matches the `scene:path` at `EditScope.out`, as mentioned in the comment.