Fix testDeleteBeforeSaveCompletesWithScriptVariables
This test checks for whether directories exist, however, the parent directory was set to only have the read flag set.  This meant that none of the exists checks could possibly do anything - they just got an exception as soon as they tried to stat the parent directory.  Arguably, since the parent directory is read only, it's redundant to check that the directories don't exist ( anything that tried to create them would fail anyway ), but by setting the execute flag, we can actually list the contents of the directory and check that specific folders don't exist.