Documentation : Fix TOC on pages containing headerlinks under figures
It appears that after migrating from ReST to MyST Sphinx now also uses the `headerlink` class for links under figures. This broke the TOC generation as it was previously assumed that every `headerlink` would have a `<H#>` parent. We now ignore headerlinks that aren't a child of a heading when generating the TOC.