Cycles ShaderNetworkAlgo : Remove `isOutput` logic
This appears to date back to a time when there was a special "output" shader type in GafferCycles, which collected surface, displacement and volume networks into a single package, and added other settings like MIS flags. This is now dealt with via a combination of individual shader assignments and the CyclesAttributes node. That change (along with all sorts of others)
 was made in 3cc31f2633704bd922ec5e65a39d5051a6f5173e.