CyclesShader : Simplify `outPlug()` handling
This seems to have been copied from OSLShader, which had an unfortunate history of creating a CompoundPlug (which is no longer a thing) for shaders with output parameters, and just a Plug for shaders without output parameters. Since we always want just a Plug, we can create it in the constructor instead. Ideally we would simplify all the Shader subclasses in this vein, and have the "out" plug be made by the base class.