BackgroundMethod : Prevent `postCall` to zombie widgets
I tried pretty hard to avoid resorting to `_qtObjectIsValid()`, but became convinced that it was the only option.

My initial attempt was to only pass `weakref( widget )` to `backgroundFunction`, but it still needs to turn that into a strong reference to call the method at some point. If that happens after the UI thread
has dropped all references to the widget, then the widget is destroyed on the background thread, which is no good because Qt objects need to be destroyed on the thread they belong in. I think this was also causing `__CurrentCall` to be destroyed from inside `backgroundFunction`, causing `backgroundFunction` to end up waiting on _itself_. This caused a shutdown crash in the new test case.