Cycles ShaderNetworkAlgo : Don't rely on `Shader::getType()`
We were using `Shader::getType()` to determine which output of the Cycles shader graph to connect to. But when loading a USD file, the type is often just the generic "shader" type, which doesn't correspond to any output. This was reported in the following gaffer-dev thread : https://groups.google.com/g/gaffer-dev/c/Sb4Opvz4i_w/m/fnGOVz7tAQAJ.
Since we already know the purpose of each network from the arguments to `convertGraph()`, we just use that information instead. This is consistent with our long-term trajectory - see related comments in `IECoreUSD::ShaderAlgo::readShaderNetwork()`.

I've also refactored `convertGraph`, so the networks are dealt with identically in a loop rather than having the code duplicated for each one.