CreateViewsUI : Adapt to new PlugValueWidget API
This also fixes a bug which allowed views to be added to read-only nodes. We were calling `self.setEnabled()` which is overridden by the PlugLayout class. The general rule is that you can manage the enabled state of your children, but not of yourself (that is your parent's responsibility). So we now set the enabled state on our button instead.