PathListingWidget : Use PathColumn `SizeMode` to control size behaviour
Columns with a sizeMode of `Strech` will be sized to take advantage of available space within the header. This resizing can occur when other columns are resized or removed, and when the geometry of the header is changed. When multiple columns are set to `Stretch`, the available space is shared between them while maintaining the relative column proportions.
If the overall width of the header is greater than the viewable region (potentially requiring a horizontal scroll bar), then we don't resize the stretchable columns. This allows a PathListingWidget to be resized to reveal more or less of the scrollable area without the stretchable columns immediately shrinking to fit the currently visible region. The streching resumes once the entire width of the header is visible, either from enlarging the viewport or reducing column widths such that scrolling is no longer necessary.

We disable Qt's `stretchLastSection` behaviour and replace it with our own, as Qt would otherwise stretch the last section and claim any available space before we have a chance to resize our columns. We replicate this behaviour when no column has a sizeMode of Stretch defined, resulting in the last column stretching to fill available space. Whether this mode is enabled or not, the last column is also automatically stretched when user resizing of other columns would result in empty space, the last column width is stored and reapplied as part of `updateColumnWidths()` in order to improve consistency across model updates and when adding and removing columns.

Calling `updateColumnWidths()` as part of `setColumns` reduces flickering in the PathListingWidget when columns are first added with Qt's default column widths and then later resized as part of the model update.