ie/installAll : Allow for selective builds
With this change it's should be possible to set specific
application/libraries to install via an argument setting that specific
version.

Note that for apps, you should use ""{APPNAME}_VERSION=", instead of
"APP_VERSION=" to specify the version to be used.

Also, once an app version is specified, it's assumed that only that app
(or other apps that have specific versions set) should be installed.

Standalone libraries are always installed.