IECoreDelight : Move module out of GafferDelight
This contains the low-level Cortex componenents that allow Gaffer to interface with 3Delight, but which don't have a dependency on Gaffer itself. The original intention was that we'd move them back to Cortex, but it has since become obvious that it's easier to manage here in the Gaffer repo. In fact, we'd really quite like to host _all_ the relevant bits of Cortex in the Gaffer repo, but that's a problem for another day.

This brings IECoreDelight/GafferDelight into line with IECoreArnold/GafferArnold.