IECoreDelight : Add display driver
This was originally provided by Cortex, and installed as a `.so` which 3Delight loaded via the `DL_DISPLAYS_PATH`. But more recent versions of 3Delight have ditched the search path and therefore can't find the `.so`. So instead, we move the code to IECoreDelight and register it directly via `DspyRegisterDriverTable`. This seems to be what the 3Delight developers recommend.

There are two potential downsides here :

- Presumably we can't write a `.nsi` file containing an `ieDisplay` and then render it directly with `renderdl`, because the driver won't be registered.
- We're no longer sharing the implementation with a potential PRMan display driver.

The alternative would be to continue to use a `.so`, but do the search path resolution ourself, and pass the full path to 3Delight. But then the resulting `.nsi` file is non-portable because it's hardcoded to the installation location (and platform) that it was written from.

Only time will tell what the right decision is here, but I'm currently swayed by the simplicity of hosting the display driver directly in IECoreDelight, and being able to remove one more component from Cortex.