Metadata : Allow concurrent access to values for the same instance
We have always provided concurrent access to values on different instances, so that we could load scripts (writing metadata) in a background thread while doing other work (potentially manipulating metadata) on a different script on the foreground thread. But we now allow multiple threads to concurrently access metadata on the _same_ node or plug.

This will make it easier to implement widgets that do work in the background and might need access to metadata as part of that. The primary motivation is to allow an asynchronous PresetsPlugValueWidget to call `NodeAlgo::currentPreset()` from a background thread, as part of #5055.