CyclesShader : Load `emission` as `cycles:surface`, not `cycles:light`
The loading as `cycles:light` originated in d6bd98b1c12dac2337686ad28d1374b3ea0cab18, which claimed that "every mesh assigned an emission shader will become a light". I don't understand that statement, because :

- An object is not a light unless it is also in the `__lights` set.
- As implemented, `Renderer::light()` is ignoring geometry anyway, so even if the object was in the lights set, the geometry would be ignored.

Now we load as `cycles:surface`, the `emission` shader can be assigned as usual to geometry, and it renders properly, as demonstrated in the modified InteractiveCyclesRenderTest.

Bafflingly, this fixes the rendering of CyclesMeshLight too! You'd expect the mesh light to need the internal shader to be assigned as `cycles:light`, but that's actually not the case, because CyclesMeshLight _doesn't_ put the object in the `__lights` set. So in fact, CyclesMeshLight just outputs regular geometry, and therefore needs a regular shader assignment too.

Fixes #5058.